First attempt at setting up a broad-vs-deep execution model.
Seems to work OK in the base case. Removes strategies, with their vestigal
remains being used for submodes of the "broad" mode, and is still pretty
hacky, e.g. some duplicated code for "loop over connections and run X",
and having to use ENV to transmit the current host info from main() to
individual operations when in "deep" mode..