Fix #7: Implement "something like roles."
Actually, what we have is a way to simulate roles.

We allow fab_hosts (and other places) to be a string rather
than a list of hosts. This string will then be used to
look up a variable that contains the real list of hosts to
connect to.

So we can have fabfiles like this:

config.servers = ['localhost', '127.0.0.1']
def prod():
  config.servers = ['...', '...']
@hosts('servers')
def deploy():
  run('stuff on $(fab_host)')

Here, we are allowed to define as many variables with lists
of hosts as we want, and we can change them any time we want.