Allow user-defined Fab recipes, in deep mode, to connect even sans operations.
In other words, you can define a method that does not use run() or sudo() or etc, and it will still make the connections and run once per host, if you're in deep mode. Useful for troubleshooting or testing; no obvious drawbacks since it only applies to user-defined functions and not internals.

May want to add an option for this at some point in case it does cause problems.