Update outputter threads to wait till threads exit, to fix occasional problems.
Was still getting intermittent cases of e.g. run() returning prematurely, while
the outputter thread was still running and pulling in data from the network. I'm not positive it's 100% fixed but it's behaving a lot better, and almost as importantly, I think conceptually this code makes a lot more sense.

Also includes extra comments in the same area.