Shuffled around contrib and added 1st draft of some file-centric stuff.
fabric.contrib.files is a near-direct import of some personal code I've been
using and building lately. Chances are it will change around a lot before
0.9-final (or may even get backed out of master and put in a feature branch in
  prep for 1.0, maybe.)

It is currently just a bunch of functions but I may try to make it more OOP if
that makes sense. May also start tying in path.py-like functionality (if that
project isn't dead, that is; site's not loading right now.)