Minor docs tweak, mostly done with fabfile docs stuff.
For some bizarro reason, the first time build_docs runs, Sphinx almost doesn't
pick up the changes to the RST file, such that it still doubles up on the
autofunctions. Running it a 2nd time without touching anything appears to make
it work better. Made sure that there was no difference between manual edits to
the file and what the new fabfile task does.

Sphinx is definitely cleaning when I tell it to, so it's not stale HTML, and it
says 11 files each time, so it doesn't appear to be disregarding the modified
file. Double check other makefile settings, maybe.