Update exception handling so unknown host rejection actually works.
Prior to this, host rejection was treated like an auth failure, giving users no
indication that it was actually due to A) unknown key and B) them having set
Fabric to reject unknown keys.

This is sort-of a reversal of
http://github.com/bitprophet/fabric/commit/7d065fbddacfe79b4fa8912a6ca523ab059fb123
in that older Paramiko releases will cause Fabric to bail out because they
return SSHException instead of AuthenticationException. Not too worried; we're
officially saying that Paramiko 1.7.4 is required at this point. If users are
using the older version, they'll get a normal Fabric 'abort' with the 'bad
password' message, which is at least a bit better than a full-on stack trace.