Backslash escaping makes it difficult to, you know. Put in backslashes.
Cannot remember why I felt they needed escaping. I really hate escaping in
general; always manage to miss one aspect of it or another. Things seem to work
OK with it taken out, and I was unable to e.g. escape dollar-signs and such in
awk commands, with it in place.