Decided not to bother with trying to "fix" autodoc+decorators issues.
Instead, will have to just break DRY a bit and manually keep up API sigs for
functions which are decorated. Not worth going down the rabbit hole which is
the alternative (e.g. I started having to do ADDITIONAL work to handle stuff
decorated by external decorators, on top of the crazy stuff I was already
doing. Just not worth it.)