Allow lazy (callable) role values in env.roledefs.
Useful for dynamically generating role host-lists, but not performing
that calculation at fabfile module level. (E.g. if you connect to a
server to grab the host list, this incurs latency for every single
invocation of "fab", even if you're not using those roles. Lazy roles
lets you defer that network connection for when you really need it.)