Initial implementation of this.
There are a few issues, some in the scp module, some in fabric
itself. Ok not fabric so much as the semantics that would require
changing and/or require changing scp.py.

In this vein I have a request in to the original author of scp.py
for more info.