Simplify this by removing the extraneous decorator
This removes the @taskdecorator as an internal decorator and does
everything within the Task.  To do so, I had to add some additional
logic to the WrappedCallableTask to make sure it looks to the outside
world like it's the original task that is being decorated.

This does break one thing within the WrappedCallableTask.  It no longer
has a `run` property, instead that is called `wrapped` to signify its
purpose in the class.