Update to handle invoked @task() and original @task
Per feedback from @bitprophet, the original tests were a little
simplistic.  This expands on those tests and fixes some issues where the
decorator would be confused if used as an invoked decorator and a
non-invoked decorator.

This may be more complex than necessary as I normally write decorators
to be one style or the other.  That said, the test cases pass, so this
is a point to start refactoring from.