Handle case where `args[0]` is not callable.
This code may go away.  Handling `args` is going to be error prone and
could lead to very bad situations.  For example, consider this code:

    def helper():
        pass

    @task(helper)
    def real_task():
        pass

In this case, `real_task` won't be wrapped, `helper` will.  I think this
edge case is enough to remove the need to handle `args`, which will
simplify the code for determining whether this was invoked.