Don't use settings() inside execute().
It breaks the technique of manipulating some env vars inside of one task
in order to affect tasks invoked after it.

While this technique is not strictly documented and is an implementation
detail, it's very useful and should not be broken until a more Pythonic
option has been added.

This use of settings() was not documented when 1.3.0 was released, so it
does not count as breaking backwards compat with 1.3.0-1.3.2.