[contrib.files] Corrected files.append regex escaping method to have consistent results with egrep used in contains
The escaping method used before was re.escape, which escapes all non alphanumeric characters. This is not compatible with many other implementations of regexp.
files.contains which uses egrep to verify the existance of the pattern was breaking when the line contained _ or < characters since it did not expect them to be escaped
Using the proposed solution of http://bugs.python.org/issue2650, only regexp metacharacters are now escaped