Adds convenience chown-ing kwargs to put and upload_template.
So that upload_template can be a one and done command, if user and/or
group kwargs are supplied to upload_template or put, the file is
chowned with the specified user or group as part of the sftp upload.

If only or the other of user or group is supplied the existing uid
or gid of the file is used when chown'ing the file.

This patch implements #257