Issue-766 Add name attribute to tasks without one.
New-style Task subclass objects which do not define a name attribute end
up with the default obj.name = 'undefined' from Task.  When the tasks
are extracted from a loaded fabfile, obj.name should be set to the name
of the instance variable, since obj.name is used internally by fabric in
some circumstances, e.g. when `execute()` is called.  This commit sets
`obj.name = name` for task objects without a name, or with an undefined
name.