FEAT: first attempt for run_winrm feature.
As using ssh seems quite ingrained in fabric code-base, and because it
is difficult to decide before-hand whether we want to use ssh or winrm,
I added a new operation winrm_run.

It supports the following:

	- buffered output (both stdout/stderr)
	- with cd context

Everything else only works by accident (in particular, if a command
expects input, it will not work ATM).