Patch input_loop not to use select
Stanley uses eventlets and monkey patches select. sys.stdin is a
blocking filesystem file descriptor in stdlib. Eventlet select
requires all sockets to be non-blocking (due to their dependency
on epoll). So instead of using select to monitor sys.stdin we use
the blocking call sys.stdin.read(1). Since sys.stdin is consumed
in a thread in fabric, this is just fine.