Add --always-exclude-hosts option.
This is exactly like --exclude-hosts, except that it is respected
by nested calls to `execute()`. So if you have a machine that is
definitely down, and you just want to skip any connection that would
normally be made to that machine, you can set `env.always_exclude_hosts`
at the very top of your supertask and know that it will be respected
all the way down through the tree of subtasks (unless of course they
explicitly reset `env.always_exclude_hosts` themselves).