Hide the default value when a validate is specified
Today asking for a basic "yes/no" question with a validate and a default value implies that the prompt display the default value into square brackets.  
For instance:

```
tagRequired = prompt('Tag this version? [Y/n]', validate=r'^[yYnN]$', default='y').lower()
```

results in:

`Tag this version? [Y/n] [y]`

which is annoying.

So my proposition is to not display the default value when a validate is specified.
Which makes sense because generally, a validate implies that you'll inform your user of the conditions of validation