Fix shell_env for win32 environments
This ought to close #1083

There were two issues with the previous implementation - basically the windows `SET` call works nothing like `export`:

1. It can only set a single value at a time (so we have to `&&` several `SET` commands together)
2. `SET` has its own special snowflake requirements for escaping the value argument - specifically:

> The characters <, >, |, &, ^ are special command shell characters, and they must be preceded by the escape character (^) or enclosed in quotation marks when used in String (for example, "StringContaining&Symbol"). If you use quotation marks to enclose a string that contains one of the special characters, the quotation marks are set as part of the environment variable value.

(Source: https://technet.microsoft.com/en-us/library/cc754250.aspx)

So enclosing quotes are taken literally and should never be escaped, nor are they necessary in order to deal with spaces (which kind of makes sense I in a way I suppose since you can only set one value at a time and the value is always the last thing in the invocation). This actually includes any trailing spaces before the `&&` so all of those had to be special cased as well. This ended up making `_prefix_env_vars` kind of messy and `shell_env_str` should probably just be composed separately in the `if win32` clause, but hey it works.

I have only tested this with Windows 7 but the technet page suggests the same semantics apply all the way back to Windows XP, so it's probably fine?