contrib.files: get rid of _expand_path() and is_win()
They're not needed. If the filename is just put into the command
string unquoted, it will be expanded by the shell.

The exception to this is that _expand_path() would join
space-separated parts (after the shell expanded each part
separately) with single spaces and then quote the whole thing
to be treated as a single filename. To keep working with
filenames with single spaces, we need to backslash-escape
spaces. As a bonus, these functions now work with two
spaces in a row, and more types of shell expansion like '?'.