Close client in case we failed to connect
The client is initialized before the try/except block, but is not
cleared on failure leaving the background threads running.

With multiple failing attempts (eg. bad key):

    while True:
        try:
            fabric.operations.run(...)
        except fabric.network.NetworkError:
            pass

this blocks all streams making paramiko to return EOFError and causing
Fabric to finish with sys.exit(0) (as if ctrl+c was pressed). Perhaps
the ctrl+c detection should be improved, but closing the client avoids
EOFError caused by this reason.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>