Dependent tasks are now executed per host.
The run() and sudo() methods on a Connection give the expectation that the
command will be run on the target host(s). However if a task is specified
as a pre or post dependency then the command is run on the machine where
the fab command was run.

This PR modifies the calls in the Executor.expand_calls() method to pass
the value of allow_hosts to any dependent tasks. The result is that instead
of the dependent tasks being executed only once on source machine they are
excuted on each target host.