default to RejectPolicy
The argument here is that it's a really bad idea to silently accept
new keys. It's not how SSH works, and it's not how Fabric *should*
work. If we'd work like SSH, we'd *prompt* the user, but the
WarningPolicy doesn't do that, it just warns.

At the very least those policies should noisily warn the user when the
host keys get changed, but that's something Paramiko must do.