default host key policy to RejectPolicy
It's a really bad idea to silently accept new keys. It's not how SSH
works, and it's not how Fabric *should* work. If we'd work like SSH,
we'd *prompt* the user, but the WarningPolicy doesn't do that, it just
warns, and doesn't prompt.

At the very least those policies should noisily warn the user when the
host keys get changed, but that's something Paramiko must do.

An alternative to this would be WarningPolicy but I think it's too
light and possibly too late: when you warn, Fabric probably has time
to run the bad stuff before the operator notices and interrupts.