do not break type decorations in fabric tasks
The invoke.task decorator has type decorations which means that
decorated functions can still be typed-checked.

But it seems fabric was not converted to that yet, so you get errors
like this with mypy --strict and, indeed, those functions are not
checked for typing.

fabric_tpa/postgresql.py:505: error: Untyped decorator makes function "backup" untyped  [misc]

The patch reuses the type signature of invoke.task. It's not actually
the exact recommendations mypy provides for that kind of stuff:

https://mypy.readthedocs.io/en/stable/generics.html#declaring-decorators

... but given the complexity of the decorator, it's not too bad.

The types are in a separate type stub file because otherwise they
break parsing in older Python releases.